backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.2.0
supabase>=2.4.0
rembg>=2.0.57
Pillow>=10.3.0
transformers>=4.40.0
torch>=2.3.0 --index-url https://download.pytorch.org/whl/cpu
replicate>=0.25.0
sse-starlette>=2.1.0
httpx>=0.27.0
boto3>=1.34.0
